sync_FL_Client/requirements_temp.txt
absl-py==0.13.0
astor==0.8.0
gast==0.4.0 #0.3.3
google-pasta==0.2.0
grpcio==1.39.0 #1.32.0
grpcio-tools==1.39.0
h5py==2.10.0 #3.1.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.1.1
numpy==1.19.5
opt-einsum==3.3.0
pandas==1.3.4
protobuf==3.10.0
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
scipy==1.3.1
six==1.15.0
tensorboard==2.4.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
Werkzeug==0.16.0
wheel==0.36.2
wrapt==1.12.1
